feat: feat_yunsun screen->template
feat_yunsun의 screen 단위를 template 단위로 바꿈

Relatedto:#6 #20
