+++ app/package.json
@@ -29,8 +29,7 @@
     "react-redux": "^9.1.1",
     "redux": "^5.0.1",
     "redux-logger": "^3.0.6",
-    "redux-saga": "^1.3.0",
-    "tailwindcss": "^3.3.2"
+    "redux-saga": "^1.3.0"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
@@ -49,6 +48,7 @@
     "prettier": "2.8.8",
     "react-devtools": "^5.0.2",
     "react-test-renderer": "18.2.0",
+    "tailwindcss": "^3.3.2",
     "typescript": "5.0.4"
   },
   "engines": {
